requirements.txt
streamlit>=1.30.0
numpy>=1.24.0
python-dotenv>=1.0.0
requests>=2.31.0
